proper usage of size_t
git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@3698 419231ca-4411-0410-aba2-fc6cf8c4f609

+++ common/exprtmpl.hh
@@ -234,7 +234,7 @@ public:
   typedef typename BlockExpression<Ex>::type BlockExpr;
   BlockExpr operator[] ( int i ) const {
     return ex[i]; }
-  int N() const { return ex.N(); }
+  size_t N() const { return ex.N(); }
   double one_norm() const { return eval_one_norm(*this); }
   double one_norm_real() const { return eval_one_norm_real(*this); }
   double two_norm() const { return sqrt(eval_two_norm2(*this)); }
@@ -255,7 +255,7 @@ public:
   typedef typename BlockType<I>::type block_type;
   typedef typename FieldType<I>::type field_type;
   //! dimension of the vector space
-  int N() const {
+  size_t N() const {
     return asImp().N();
   }
   double one_norm() const { return eval_one_norm(*this); }
@@ -323,40 +323,40 @@ public:
     Dune::dvverb << INDENT << "Assign Vector from field_type\n";
 #endif
     ++INDENT;
-    for (int i=0; i<N(); ++i) { asImp()[i] = x; }
+    for (size_t i=0; i<N(); ++i) { asImp()[i] = x; }
     --INDENT;
     return asImp();
   }
   template <class E> Vector<I>& operator+=(const Expression<E>& x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] += x[i];
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] += x[i];
     return asImp();
   }
   template <class V> Vector<I>& operator+=(const Vector<V>& x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] += x[i];
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] += x[i];
     return asImp();
   }
   template <class E> Vector<I>& operator-=(const Expression<E>& x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] -= x[i];
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] -= x[i];
     return asImp();
   }
   template <class V> Vector<I>& operator-=(const Vector<V>& x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] -= x[i];
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] -= x[i];
     return asImp();
   }
   Vector<I>& operator+=(field_type x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] += x;
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] += x;
     return asImp();
   }
   Vector<I>& operator-=(field_type x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] -= x;
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] -= x;
     return asImp();
   }
   Vector<I>& operator*=(field_type x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] *= x;
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] *= x;
     return asImp();
   }
   Vector<I>& operator/=(field_type x) {
-    for (int i=0; i < asImp().N(); i++) asImp()[i] /= x;
+    for (size_t i=0; i < asImp().N(); i++) asImp()[i] /= x;
     return asImp();
   }
 private:
@@ -379,7 +379,7 @@ public:
 #endif
     return BlockExprImp(v[i]);
   }
-  int N() const { return v.N(); };
+  size_t N() const { return v.N(); };
   double one_norm() const { return eval_one_norm(*this); }
   double one_norm_real() const { return eval_one_norm_real(*this); }
   double two_norm() const { return sqrt(eval_two_norm2(*this)); }
@@ -467,7 +467,7 @@ public:
   typedef typename RowType<I>::type row_type;
   typedef typename FieldType<I>::type field_type;
   //! dimension of the vector space
-  int N() const {
+  size_t N() const {
     return asImp().N();
   }
   int M() const {
@@ -713,7 +713,7 @@ public:
     M[MyDepth<B,Mat>::value] = i;
     return SubMatrixMulVector(A,v,M,*this);
   }
-  int N() const { return -1; };//r.begin()->N(); }
+  size_t N() const { return -1; };//r.begin()->N(); }
   const ParentBlockType & parent;  
 private:
   mutable int* M;
@@ -740,7 +740,7 @@ public:
     M[0] = i;
     return SubMatrixMulVector(A,v,M,*this);
   }
-  int N() const { return -1; }; // { parent.begin().N(); }
+  size_t N() const { return -1; }; // { parent.begin().N(); }
 private:
   mutable int* M;
   const Mat & A;
@@ -771,7 +771,7 @@ public:
     }
     return x;
   }
-  int N() const { return iN; };
+  size_t N() const { return iN; };
   const ParentBlockType & parent;
 private:
   mutable int* M;
@@ -798,7 +798,7 @@ public:
     }
     return x;
   }
-  int N() const { return iN; };
+  size_t N() const { return iN; };
 private:
   const Mat & A;
   const Vec & v;
@@ -926,7 +926,7 @@ struct FieldType< ExprTmpl::Matrix<I> >
   typename FieldType<A>::type val=0; \
   Dune::dvverb << INDENT << "Infinity Norm of Expression\n"; \
   ++INDENT; \
-  for (int i=0; i<a.N(); ++i) { val += eval_one_norm(a[i]); } \
+  for (size_t i=0; i<a.N(); ++i) { val += eval_one_norm(a[i]); } \
   --INDENT; \
   return val; \
 }
@@ -954,7 +954,7 @@ one_norm (const A & a)
   typename FieldType<A>::type val=0; \
   Dune::dvverb << INDENT << "Infinity Norm of Expression\n"; \
   ++INDENT; \
-  for (int i=0; i<a.N(); ++i) { val += eval_one_norm_real(a[i]); } \
+  for (size_t i=0; i<a.N(); ++i) { val += eval_one_norm_real(a[i]); } \
   --INDENT; \
   return val; \
 }
@@ -991,7 +991,7 @@ two_norm (const A & a)
   typename FieldType<A>::type val=0; \
   Dune::dvverb << INDENT << "Infinity Norm of Expression\n"; \
   ++INDENT; \
-  for (int i=0; i<a.N(); ++i) { val += eval_two_norm2(a[i]); } \
+  for (size_t i=0; i<a.N(); ++i) { val += eval_two_norm2(a[i]); } \
   --INDENT; \
   return val; \
 }
@@ -1018,7 +1018,7 @@ two_norm2 (const A & a)
   typename FieldType<A>::type val=0; \
   Dune::dvverb << INDENT << "Infinity Norm of Expression\n"; \
   ++INDENT; \
-  for (int i=0; i<a.N(); ++i) { val = std::max(val,eval_infinity_norm(a[i])); } \
+  for (size_t i=0; i<a.N(); ++i) { val = std::max(val,eval_infinity_norm(a[i])); } \
   --INDENT; \
   return val; \
 }
@@ -1039,7 +1039,7 @@ infinity_norm (const A & a)
   typename FieldType<A>::type val=0; \
   Dune::dvverb << INDENT << "Infinity Norm of Expression\n"; \
   ++INDENT; \
-  for (int i=0; i<a.N(); ++i) { val = std::max(val,eval_infinity_norm(a[i])); } \
+  for (size_t i=0; i<a.N(); ++i) { val = std::max(val,eval_infinity_norm(a[i])); } \
   --INDENT; \
   return val; \
 }
@@ -1070,7 +1070,7 @@ operator * (const Vector<A> & a, const Vector<A> & b)
 {
   assert(a.N() == b.N());
   typename FieldType<A>::type x = 0;
-  for (int i=0; i<a.N(); i++)
+  for (size_t i=0; i<a.N(); i++)
     x = a[i] * b[i];
   return x;
 }
@@ -1083,7 +1083,7 @@ operator * (const Vector<A> & a, const Expression<B> & b)
   IsTrue< SameType<FieldType<A>,FieldType<B> >::value == true >::yes();
   assert(a.N() == b.N());
   typename FieldType<A>::type x = 0;
-  for (int i=0; i<a.N(); i++)
+  for (size_t i=0; i<a.N(); i++)
     x = a[i] * b[i];
   return x;
 }
@@ -1096,7 +1096,7 @@ operator * (const Expression<A> & a, const Vector<B> & b)
   IsTrue< SameType<FieldType<A>,FieldType<B> >::value == true >::yes();
   assert(a.N() == b.N());
   typename FieldType<A>::type x = 0;
-  for (int i=0; i<a.N(); i++)
+  for (size_t i=0; i<a.N(); i++)
     x = a[i] * b[i];
   return x;
 }
